Rename diagnostics prefix from ST  to AVTR
This removes the last remainings from the old naming.
