+++ src/Avatar.UnitTests/Avatar.UnitTests.csproj
@@ -56,9 +56,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Include="CodeAnalysis\ST*\**\*.NoBuild.cs" CopyToOutputDirectory="PreserveNewest" />
-    <Compile Remove="CodeAnalysis\ST*\**\*.NoBuild.cs" />
-    <Compile Update="CodeAnalysis\ST*\**\*.cs" CopyToOutputDirectory="PreserveNewest" />
+    <None Include="CodeAnalysis\AVTR*\**\*.NoBuild.cs" CopyToOutputDirectory="PreserveNewest" />
+    <Compile Remove="CodeAnalysis\AVTR*\**\*.NoBuild.cs" />
+    <Compile Update="CodeAnalysis\AVTR*\**\*.cs" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
 
   <ItemGroup>
@@ -82,8 +82,8 @@
       <ThisAssemblySourcesDirectory>$(MSBuildProjectDirectory)\$(OutputPath)</ThisAssemblySourcesDirectory>
     </PropertyGroup>
     <ItemGroup>
-      <FileConstant Include="@(Compile)" Condition="$([MSBuild]::ValueOrDefault('%(RelativeDir)', '').StartsWith('CodeAnalysis\ST'))" />
-      <FileConstant Include="@(Compile)" Condition="$([MSBuild]::ValueOrDefault('%(RelativeDir)', '').StartsWith('CodeAnalysis/ST'))" />
+      <FileConstant Include="@(Compile)" Condition="$([MSBuild]::ValueOrDefault('%(RelativeDir)', '').StartsWith('CodeAnalysis\AVTR'))" />
+      <FileConstant Include="@(Compile)" Condition="$([MSBuild]::ValueOrDefault('%(RelativeDir)', '').StartsWith('CodeAnalysis/AVTR'))" />
       <FileConstant Include="@(Compile)" Condition="$([MSBuild]::ValueOrDefault('%(RelativeDir)', '').StartsWith('Scenarios'))" />
     </ItemGroup>
   </Target>
