+++ src/Avatar.UnitTests/CodeAnalysis/AVTR001/Diagnostic/PublicClass.cs
@@ -1,14 +1,16 @@
 ﻿using System;
 
-namespace Avatars.UnitTests.CodeAnalysis.ST004.Diagnostic
+namespace Avatars.UnitTests.CodeAnalysis.AVTR001.Diagnostic
 {
     public partial class MyClass
     {
         public MyClass()
         {
-            var avatar = Avatar.Of<PlatformID>();
+            var avatar = Avatar.Of<IDisposable, BaseType>();
 
             Console.WriteLine(avatar);
         }
     }
+
+    public class BaseType { }
 }
