+++ src/Avatar.UnitTests/CodeAnalysis/AVTR001_BaseTypeNotFist.cs
@@ -6,18 +6,18 @@ using Xunit;
 
 namespace Avatars.UnitTests
 {
-    public class ST002_DuplicateBaseType : DiagnosticVerifier
+    public class AVTR001_BaseTypeNotFist : DiagnosticVerifier
     {
         protected override DiagnosticAnalyzer? GetCSharpDiagnosticAnalyzer() => new ValidateTypesAnalyzer();
 
         [Theory]
-        [InlineData(ThisAssembly.Constants.CodeAnalysis.ST002.Diagnostic.PublicClass, 9, 26)]
+        [InlineData(ThisAssembly.Constants.CodeAnalysis.AVTR001.Diagnostic.PublicClass, 9, 26)]
         public void Verify_Diagnostic(string path, int line, int column)
         {
             var expected = new DiagnosticResult
             {
-                Id = AvatarDiagnostics.DuplicateBaseType.Id,
-                Message = Resources.DuplicateBaseType_Message,
+                Id = AvatarDiagnostics.BaseTypeNotFirst.Id,
+                Message = string.Format(Resources.BaseTypeNotFirst_Message, "BaseType"),
                 Severity = DiagnosticSeverity.Error,
                 Locations = new[] {
                     new DiagnosticResultLocation("Test0.cs", line, column)
