+++ src/Avatar.UnitTests/CodeAnalysis/AVTR004/Diagnostic/PublicClass.cs
@@ -1,16 +1,14 @@
 ﻿using System;
 
-namespace Avatars.UnitTests.CodeAnalysis.ST003.Diagnostic
+namespace Avatars.UnitTests.CodeAnalysis.AVTR004.Diagnostic
 {
     public partial class MyClass
     {
         public MyClass()
         {
-            var avatar = Avatar.Of<BaseType>();
+            var avatar = Avatar.Of<PlatformID>();
 
             Console.WriteLine(avatar);
         }
     }
-
-    public sealed class BaseType { }
 }
