+++ src/Avatar.UnitTests/CodeAnalysis/AVTR999/Diagnostic/PublicClass.cs
@@ -1,5 +1,5 @@
 ﻿#nullable disable
-namespace Avatars.UnitTests.CodeAnalysis.ST999.Diagnostic
+namespace Avatars.UnitTests.CodeAnalysis.AVTR999.Diagnostic
 {
     public partial class MyClass
     {
