+++ src/Avatar.UnitTests/CodeAnalysis/AVTR999_OverridableMembers.cs
@@ -1,26 +1,21 @@
 ﻿using System.IO;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Xunit;
 
 namespace Avatars.UnitTests
 {
-    public class ST999_OverrideAllMembers : CodeFixVerifier
+    public class AVTR999_OverridableMembers : DiagnosticVerifier
     {
         protected override DiagnosticAnalyzer? GetCSharpDiagnosticAnalyzer() => new OverridableMembersAnalyzer();
 
-        protected override CodeFixProvider? GetCSharpCodeFixProvider() => new OverrideAllMembersCodeFix();
-
         [Theory]
-        [InlineData(
-            ThisAssembly.Constants.CodeAnalysis.ST999.Diagnostic.PublicClass,
-            ThisAssembly.Constants.CodeAnalysis.ST999.Diagnostic.PublicClassFix, 4, 26)]
-        public void Verify_Diagnostic(string path, string fix, int line, int column)
+        [InlineData(ThisAssembly.Constants.CodeAnalysis.AVTR999.Diagnostic.PublicClass, 4, 26)]
+        public void Verify_Diagnostic(string path, int line, int column)
         {
             var expected = new DiagnosticResult
             {
-                Id = nameof(ThisAssembly.Constants.CodeAnalysis.ST999),
+                Id = "AVTR999",
                 Message = nameof(OverridableMembersAnalyzer),
                 Severity = DiagnosticSeverity.Hidden,
                 Locations = new[] {
@@ -31,10 +26,6 @@ namespace Avatars.UnitTests
             VerifyCSharpDiagnostic(
                 File.ReadAllText(path),
                 expected);
-
-            VerifyCSharpFix(
-                File.ReadAllText(path),
-                File.ReadAllText(fix));
         }
     }
 }
