+++ src/Avatar.UnitTests/CodeAnalysis/AVTR999_OverrideAllMembers.cs
@@ -1,21 +1,26 @@
 ﻿using System.IO;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Xunit;
 
 namespace Avatars.UnitTests
 {
-    public class ST999_OverridableMembers : DiagnosticVerifier
+    public class AVTR999_OverrideAllMembers : CodeFixVerifier
     {
         protected override DiagnosticAnalyzer? GetCSharpDiagnosticAnalyzer() => new OverridableMembersAnalyzer();
 
+        protected override CodeFixProvider? GetCSharpCodeFixProvider() => new OverrideAllMembersCodeFix();
+
         [Theory]
-        [InlineData(ThisAssembly.Constants.CodeAnalysis.ST999.Diagnostic.PublicClass, 4, 26)]
-        public void Verify_Diagnostic(string path, int line, int column)
+        [InlineData(
+            ThisAssembly.Constants.CodeAnalysis.AVTR999.Diagnostic.PublicClass,
+            ThisAssembly.Constants.CodeAnalysis.AVTR999.Diagnostic.PublicClassFix, 4, 26)]
+        public void Verify_Diagnostic(string path, string fix, int line, int column)
         {
             var expected = new DiagnosticResult
             {
-                Id = "ST999",
+                Id = nameof(ThisAssembly.Constants.CodeAnalysis.AVTR999),
                 Message = nameof(OverridableMembersAnalyzer),
                 Severity = DiagnosticSeverity.Hidden,
                 Locations = new[] {
@@ -26,6 +31,10 @@ namespace Avatars.UnitTests
             VerifyCSharpDiagnostic(
                 File.ReadAllText(path),
                 expected);
+
+            VerifyCSharpFix(
+                File.ReadAllText(path),
+                File.ReadAllText(fix));
         }
     }
 }
